pyproject.toml
[project]
name = "experiment-with-openai-oss-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "openai>=1.99.9",
    "openai-agents[litellm,viz]>=0.2.6",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]
